Fix dose data type in TS import from pattern and with angular range

+++ tomo/protocols/protocol_ts_import.py
@@ -748,7 +748,7 @@ class ProtImportTsBase(ProtTomoImportFiles):
 
             for i, a in enumerate(angles):
                 order = i + 1 if not tiltorders else tiltorders[i]
-                dose = doses[i] if doses else int(self.dosePerFrame.get()) * order
+                dose = doses[i] if doses else float(self.dosePerFrame.get()) * order
                 fileList.append(((i + 1, file), order, a, dose))
 
         addFunc = _addOne if self._anglesInPattern() else _addMany
